Sportan-backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sportan-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "email-validator>=2.2.0",
    "fastapi>=0.122.0",
    "openai>=1.51.0",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.44",
    "supabase>=2.24.0",
    "uvicorn>=0.38.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "ruff>=0.14.7",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["app"]
